\name{lambda_matrix}
\alias{lambda_matrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Lambda estimates using demographic data in a matrix approach}
\description{
A calculation of the finite rate of increase (lambda) using a simplied population projection matrix approach based on basic demographic data.

The model based on demographic data: age of first basic reproduction, number of female offspring per female, maximum age of reproduction, and  two pieces of empirical survival data: 1) the juvenile survival rate through to adulthood, and 2) annual survival rate for adults.
}
\usage{
lambda_matrix(r.stage, b, stages, terminal)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{r.stage}{Stage in which first reproduction occurs}
  \item{b}{Number of female offspring per female}
  \item{stages}{Number of distinct survival/fecundity stages}
  \item{S}{Vector of stage specific survival rates}
  \item{terminal}{Logical - whether all adults die after reaching final stage}
  }

\details{
If the stages of matrix are set to represent age and there is obligate death of adults (terminal = TRUE) then the finite rate of increase should be equivalent to that generated by the lambda_PM function ( equivalent to the "PM" equiation of Slade et al (1998)
}
\value{
Returns an estimate of the finite rate of increase for a given population based on the juvenile survival rate through to adulthood, and the adult annual survival rate.
}
\references{
Slade, N. A., Gomulkiewicz, R., & Alexander, H. M. (1998). Alternatives to Robinson and Redford's method of assessing overharvest from incomplete demographic data. Conservation Biology, 12(1), 148-155.
}
\author{
Nathan Whitmore
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
# e.g. life history data from Ochotona princeps (Slade et al. 1998)
# in which each stage is equivalent to a year
# such that r.stage is equivalent to age of first reproduction
# and number of stages is the maximum age of reproduction

lambda_matrix(
  r.stage = 1,  # stage of first reproduction
  b = 3.25,  # number of female offspring per female
  stages = 6 , # number of stages
  S = c(0.11,rep(0.7,5)),  # survival rate vector
  terminal = TRUE
)
}
